スタンプダイアログのchildrenを渡しているコンポーネントを削除する
